Introduce a 'double fixed' data type abstraction
The 'double fixed' value type is a fixed point data type implemented as
two signed 32 bit integers. It is intended to be sent over the wire and
used when wl_fixed_t is not detailed enough.

Two helper functions are introduced: wl_double_fixed_to_double and
wl_double_fixed_from_double that can be used for conversion.

The type is implemented by separating the integral part from the
fractional part, storing them in individual signed 32 bit integers.

Converting to a double could be implemented with:

	double = integral + fraction / 2147483648

Converting to a double could be implemented with:

	integral = integral part of double
	fraction = fractional part of double * 2147483648

The implementation provided in this patch is slightly faster than the
above methods, but assumes that the IEEE 754 double-precision binary
floating-point format is used internally. It is based on the wl_fixed_t
conversion functions from wayland-util.h.

Author: Jonas Ådahl <jadahl@gmail.com>
Signed-off-by: Manuel Bachmann <manuel.bachmann@iot.bzh>